\name{McCrackenData}
\docType{data}
\alias{McCrackenData}
\alias{PiBounds}
\title{Critical values for oos-t test from McCracken (2007)}
\description{
  This data set reproduces tables 1, 2, and 3 from McCracken (2007).
  The function \code{mccracken_criticalvalue} can be used to retrieve
  specific values.
}

\usage{McCrackenData}
\seealso{\code{\link{mccracken_criticalvalue}}}

\format{A list of arrays.}
\references{
  McCracken, M. 2007, Asymptotics for out of sample tests of Granger
  causality.  \emph{Journal of Econometrics}, \bold{140}(2): 719-752.
}
\keyword{datasets}