\name{extract.target}
\alias{extract.target}
\alias{extract.predictors}
\title{Convenience function to extract data from a model}
\description{Convenience function to extract data from a model}
\usage{
extract.target(model, dataset)
extract.predictors(model, dataset)
}
\arguments{
  \item{model}{}
  \item{dataset}{}
}
\details{text}
\value{text}
\references{
  Calhoun, G. 2011, Documentation appendix: An asymptotically normal
  out-of-sample test of equal predictive accuracy for nested models.
  Unpublished manuscript.
}
\author{Gray Calhoun <gcalhoun@iastate.edu>}

\examples{
##---- Should be DIRECTLY executable !! ----
}

