\name{stepm}
\alias{stepm}
\title{Romano and Wolf's (2005) StepM}
\description{Implements Romano and Wolf's StepM, a stepdown procedure
  for multiple hypothesis testing.}}
\usage{
stepm(teststatistics, bootmatrix, level)
}

\arguments{
  \item{teststatistics}{}
  \item{bootmatrix}{}
  \item{level}{}
}
\details{text}
\value{text}
\references{
  Calhoun, G. 2011, Documentation appendix: An asymptotically normal
  out-of-sample test of equal predictive accuracy for nested models.
  Unpublished manuscript.
}
\author{Gray Calhoun <gcalhoun@iastate.edu>}

\seealso{\code{\link{boot}}}
\examples{
##---- Should be DIRECTLY executable !! ----
}

